refactor(governance): move canister_state into main lib
