+++ rs/nervous_system/time_helpers/BUILD.bazel
@@ -1,0 +1,24 @@
+load("@rules_rust//rust:defs.bzl", "rust_doc_test", "rust_library", "rust_test_suite")
+
+package(default_visibility = ["//visibility:public"])
+
+DEPENDENCIES = [
+    # Keep sorted.
+    "@crate_index//:ic-cdk",
+]
+
+DEV_DEPENDENCIES = [
+]
+
+LIB_SRCS = glob(
+    ["src/**"],
+    exclude = ["**/*tests.rs"],
+)
+
+rust_library(
+    name = "time_helpers",
+    srcs = LIB_SRCS,
+    crate_name = "ic_nervous_system_time_helpers",
+    version = "0.0.1",
+    deps = DEPENDENCIES,
+)
