+++ rs/nns/governance/BUILD.bazel
@@ -45,6 +45,7 @@ DEPENDENCIES = [
     "//rs/crypto/sha2",
     "//rs/ledger_suite/common/ledger_core",
     "//rs/ledger_suite/icp:icp_ledger",
+    "//rs/nervous_system/canisters",
     "//rs/nervous_system/clients",
     "//rs/nervous_system/common",
     "//rs/nervous_system/governance",
@@ -55,6 +56,7 @@ DEPENDENCIES = [
     "//rs/nervous_system/root",
     "//rs/nervous_system/runtime",
     "//rs/nervous_system/temporary",
+    "//rs/nervous_system/time_helpers",
     "//rs/nns/cmc",
     "//rs/nns/common",
     "//rs/nns/constants",
@@ -242,7 +244,6 @@ rust_canister(
     deps = DEPENDENCIES + [
         ":build_script",
         ":governance",
-        "//rs/nervous_system/canisters",
     ],
 )
 
@@ -262,7 +263,6 @@ rust_canister(
     deps = DEPENDENCIES + [
         ":build_script",
         ":governance--test_feature",
-        "//rs/nervous_system/canisters",
         "//rs/tla_instrumentation",
     ],
 )
@@ -297,7 +297,6 @@ rust_test(
     deps = DEPENDENCIES + DEV_DEPENDENCIES + [
         ":build_script",
         ":governance",
-        "//rs/nervous_system/canisters",
     ],
 )
 
@@ -311,7 +310,6 @@ rust_test(
     deps = DEPENDENCIES + DEV_DEPENDENCIES + [
         ":build_script",
         ":governance--test_feature",
-        "//rs/nervous_system/canisters",
     ],
 )
 
@@ -333,7 +331,6 @@ rust_test(
     proc_macro_deps = MACRO_DEPENDENCIES + MACRO_DEV_DEPENDENCIES,
     deps = DEPENDENCIES + DEV_DEPENDENCIES + [
         ":build_script",
-        "//rs/nervous_system/canisters",
     ],
 )
