+++ rs/nns/governance/canister/canister.rs
@@ -1,5 +1,4 @@
-use async_trait::async_trait;
-use candid::{candid_method, Decode, Encode};
+use candid::candid_method;
 use ic_base_types::{CanisterId, PrincipalId};
 use ic_canisters_http_types::{HttpRequest, HttpResponse, HttpResponseBuilder};
 use ic_cdk::{
@@ -13,6 +12,7 @@ use ic_nervous_system_common::{
     serve_metrics,
 };
 use ic_nervous_system_runtime::{CdkRuntime, Runtime};
+use ic_nervous_system_time_helpers::now_seconds;
 use ic_nns_common::{
     access_control::{check_caller_is_gtc, check_caller_is_ledger},
     pb::v1::{NeuronId as NeuronIdProto, ProposalId as ProposalIdProto},
@@ -20,7 +20,9 @@ use ic_nns_common::{
 };
 use ic_nns_constants::LEDGER_CANISTER_ID;
 use ic_nns_governance::{
-    decoder_config, encode_metrics,
+    canister_state::CanisterEnv,
+    canister_state::{governance, governance_mut, set_governance},
+    encode_metrics,
     governance::{Environment, Governance, HeapGrowthPotential, RngError, TimeWarp as GovTimeWarp},
     is_prune_following_enabled,
     neuron_data_validation::NeuronDataValidationSummary,
@@ -50,22 +52,15 @@ use ic_nns_governance_api::{
         SettleCommunityFundParticipation, SettleNeuronsFundParticipationRequest,
         SettleNeuronsFundParticipationResponse, UpdateNodeProvider, Vote,
     },
-    subnet_rental::{SubnetRentalProposalPayload, SubnetRentalRequest},
 };
-use ic_sns_wasm::pb::v1::{AddWasmRequest, SnsWasm};
 use prost::Message;
 use rand::{RngCore, SeedableRng};
 use rand_chacha::ChaCha20Rng;
-use std::{
-    boxed::Box,
-    cell::RefCell,
-    ops::Bound,
-    str::FromStr,
-    time::{Duration, SystemTime},
-};
+use std::{boxed::Box, ops::Bound, time::Duration};
 
 #[cfg(not(feature = "tla"))]
 use ic_nervous_system_canisters::ledger::IcpLedgerCanister;
+use ic_nns_governance::canister_state::with_governance;
 
 #[cfg(feature = "tla")]
 mod tla_ledger;
@@ -79,81 +74,6 @@ use tla_ledger::LoggingIcpLedgerCanister as IcpLedgerCanister;
 const WASM_PAGES_RESERVED_FOR_UPGRADES_MEMORY: u64 = 65_536;
 
 pub(crate) const LOG_PREFIX: &str = "[Governance] ";
-
-thread_local! {
-    static GOVERNANCE: RefCell<Governance> = RefCell::new(Governance::new_uninitialized(
-        Box::new(CanisterEnv::new()),
-        Box::new(IcpLedgerCanister::<CdkRuntime>::new(LEDGER_CANISTER_ID)),
-        Box::new(CMCCanister::<CdkRuntime>::new()),
-    ));
-}
-
-/*
-Recommendations for Using `unsafe` in the Governance canister:
-
-The state of governance is captured in a mutable static global variable to allow for
-concurrent mutable access and modification of state in the NNS Governance canister. Due
-to safety checks in Rust, accessing the static variable must be done in an unsafe block.
-While this is generally an unsafe practice in normal Rust code, due to the message model of
-the Internet Computer, only one instance of the state is ever accessed at once. The following
-are best practices for making use of the unsafe block:
-
-1. Initialization First:
-    - Always ensure the global state (e.g., `GOVERNANCE`) has been initialized before access.
-      Typically, this initialization occurs in `canister_init` or `canister_post_upgrade`.
-
-2. Understanding
-    - Lifetimes in Runtime Context: When working with asynchronous functions that use mutable
-      references to Governance pay close attention to the different runtimes the code may run in:
-        - In unit tests, all futures are immediately ready. Mutating a `'static` ref is still
-          valid since futures resolve instantly, but is an abuse of the rules in Rust.
-        - In mainnet, "self" refers to the `GOVERNANCE` static variable, which is initialized
-          once in functions like `canister_init` or `canister_post_upgrade`.
-
-3. Lifetime Assurances:
-    - In a `Drop` implementation that takes mutable references of `self`, the scope of any
-      `Governance` method ensures `&self` remains alive since Governance is always
-      initialized immediately after an upgrade in the post upgrade hook. Additionally,
-      since upgrades cannot happen during an asynchronous call (the upgrade waits for
-      all open-call-contexts to be closed), Governance will never be un-initialized
-      when an async method returns. De-referencing is acceptable in this context. For
-      instance, it's always safe when a `LedgerUpdateLock` goes out of scope,
-      but requires an `unsafe` block.
-
-4. Safety Checks Inside Unsafe:
-    - Although a block is marked `unsafe`, internal verifications are still essential. For
-      instance, `unlock_neuron` within the `Drop` implementation of `LedgerUpdateLock`
-      confirms the lock's existence despite being inside an unsafe context.
-
-5. Modifying references across and await:
-    - Since the CDK will put local variables on the stack, accessing a reference across an
-      await is not advised. It is best practice to reacquire a reference to the state after
-      an async call.
-*/
-/// Returns an immutable reference to the global state.
-///
-/// This should only be called once the global state has been initialized, which
-/// happens in `canister_init` or `canister_post_upgrade`.
-fn governance() -> &'static Governance {
-    unsafe { &*GOVERNANCE.with(|g| g.as_ptr()) }
-}
-
-/// Returns a mutable reference to the global state.
-///
-/// This should only be called once the global state has been initialized, which
-/// happens in `canister_init` or `canister_post_upgrade`.
-fn governance_mut() -> &'static mut Governance {
-    unsafe { &mut *GOVERNANCE.with(|g| g.as_ptr()) }
-}
-
-// Sets governance global state to the given object.
-fn set_governance(gov: Governance) {
-    GOVERNANCE.set(gov);
-
-    governance()
-        .validate()
-        .expect("Error initializing the governance canister.");
-}
 
 fn schedule_timers() {
     schedule_seeding(Duration::from_nanos(0));
@@ -278,171 +198,6 @@ fn schedule_vote_processing() {
     ic_cdk_timers::set_timer_interval(VOTE_PROCESSING_INTERVAL, || {
         spawn(governance_mut().process_voting_state_machines());
     });
-}
-
-struct CanisterEnv {
-    rng: Option<ChaCha20Rng>,
-    time_warp: GovTimeWarp,
-}
-
-fn now_nanoseconds() -> u64 {
-    if cfg!(target_arch = "wasm32") {
-        ic_cdk::api::time()
-    } else {
-        SystemTime::now()
-            .duration_since(SystemTime::UNIX_EPOCH)
-            .expect("Failed to get time since epoch")
-            .as_nanos()
-            .try_into()
-            .expect("Failed to convert time to u64")
-    }
-}
-
-fn now_seconds() -> u64 {
-    Duration::from_nanos(now_nanoseconds()).as_secs()
-}
-
-impl CanisterEnv {
-    fn new() -> Self {
-        CanisterEnv {
-            rng: None,
-            time_warp: GovTimeWarp { delta_s: 0 },
-        }
-    }
-}
-
-#[async_trait]
-impl Environment for CanisterEnv {
-    fn now(&self) -> u64 {
-        self.time_warp.apply(now_seconds())
-    }
-
-    fn set_time_warp(&mut self, new_time_warp: GovTimeWarp) {
-        self.time_warp = new_time_warp;
-    }
-
-    fn random_u64(&mut self) -> Result<u64, RngError> {
-        match self.rng.as_mut() {
-            Some(rand) => Ok(rand.next_u64()),
-            None => Err(RngError::RngNotInitialized),
-        }
-    }
-
-    fn random_byte_array(&mut self) -> Result<[u8; 32], RngError> {
-        match self.rng.as_mut() {
-            Some(rand) => {
-                let mut bytes = [0u8; 32];
-                rand.fill_bytes(&mut bytes);
-                Ok(bytes)
-            }
-            None => Err(RngError::RngNotInitialized),
-        }
-    }
-
-    fn seed_rng(&mut self, seed: [u8; 32]) {
-        self.rng.replace(ChaCha20Rng::from_seed(seed));
-    }
-
-    fn get_rng_seed(&self) -> Option<[u8; 32]> {
-        self.rng.as_ref().map(|rng| rng.get_seed())
-    }
-
-    fn execute_nns_function(
-        &self,
-        proposal_id: u64,
-        update: &gov_pb::ExecuteNnsFunction,
-    ) -> Result<(), gov_pb::GovernanceError> {
-        // use internal types, as this API is used in core
-        use gov_pb::{governance_error::ErrorType, GovernanceError, NnsFunction};
-
-        let mt = NnsFunction::try_from(update.nns_function).map_err(|_|
-            // No update type specified.
-            GovernanceError::new(ErrorType::PreconditionFailed))?;
-
-        let reply = move || {
-            governance_mut().set_proposal_execution_status(proposal_id, Ok(()));
-        };
-        let reject = move |(code, msg): (i32, String)| {
-            let mut msg = msg;
-            // There's no guarantee that the reject response is a string of character, and
-            // it can also be potential large. Propagating error information
-            // here is on a best-effort basis.
-            const MAX_REJECT_MSG_SIZE: usize = 10000;
-            if msg.len() > MAX_REJECT_MSG_SIZE {
-                msg = "(truncated error message) "
-                    .to_string()
-                    .chars()
-                    .chain(
-                        msg.char_indices()
-                            .take_while(|(pos, _)| *pos < MAX_REJECT_MSG_SIZE)
-                            .map(|(_, char)| char),
-                    )
-                    .collect();
-            }
-
-            governance_mut().set_proposal_execution_status(
-                proposal_id,
-                Err(GovernanceError::new_with_message(
-                    ErrorType::External,
-                    format!(
-                        "Error executing ExecuteNnsFunction proposal. Error Code: {}. Rejection message: {}",
-                        code,
-                        msg
-                    ),
-                )),
-            );
-        };
-        let (canister_id, method) = mt.canister_and_function()?;
-        let method = method.to_owned();
-        let proposal_timestamp_seconds = governance()
-            .get_proposal_data(ProposalId(proposal_id))
-            .map(|data| data.proposal_timestamp_seconds)
-            .ok_or(GovernanceError::new(ErrorType::PreconditionFailed))?;
-        let effective_payload = get_effective_payload(
-            mt,
-            update.payload.clone(),
-            proposal_id,
-            proposal_timestamp_seconds,
-        )?;
-
-        spawn(async move {
-            match CdkRuntime::call_bytes_with_cleanup(canister_id, &method, &effective_payload)
-                .await
-            {
-                Ok(_) => reply(),
-                Err(e) => reject(e),
-            }
-        });
-
-        Ok(())
-    }
-
-    async fn call_canister_method(
-        &self,
-        target: CanisterId,
-        method_name: &str,
-        request: Vec<u8>,
-    ) -> Result<Vec<u8>, (Option<i32>, String)> {
-        CdkRuntime::call_bytes_with_cleanup(target, method_name, &request)
-            .await
-            .map_err(|(code, msg)| (Some(code), msg))
-    }
-
-    #[cfg(target_arch = "wasm32")]
-    fn heap_growth_potential(&self) -> HeapGrowthPotential {
-        if core::arch::wasm32::memory_size(0)
-            < ic_nns_governance::governance::HEAP_SIZE_SOFT_LIMIT_IN_WASM32_PAGES
-        {
-            HeapGrowthPotential::NoIssue
-        } else {
-            HeapGrowthPotential::LimitedAvailability
-        }
-    }
-
-    #[cfg(not(target_arch = "wasm32"))]
-    fn heap_growth_potential(&self) -> HeapGrowthPotential {
-        unimplemented!("CanisterEnv can only be used with wasm32 environment.");
-    }
 }
 
 // We expect PrincipalId for all methods, but ic_cdk returns candid::Principal, so we need to
@@ -728,7 +483,7 @@ fn get_neuron_info_by_id_or_subaccount(
 #[query]
 fn get_proposal_info(id: ProposalId) -> Option<ProposalInfo> {
     debug_log("get_proposal_info");
-    GOVERNANCE.with_borrow(|governance| governance.get_proposal_info(&caller(), id))
+    with_governance(|governance| governance.get_proposal_info(&caller(), id))
 }
 
 #[query]
@@ -744,13 +499,13 @@ fn get_neurons_fund_audit_info(
 #[query]
 fn get_pending_proposals() -> Vec<ProposalInfo> {
     debug_log("get_pending_proposals");
-    GOVERNANCE.with_borrow(|governance| governance.get_pending_proposals(&caller()))
+    with_governance(|governance| governance.get_pending_proposals(&caller()))
 }
 
 #[query]
 fn list_proposals(req: ListProposalInfo) -> ListProposalInfoResponse {
     debug_log("list_proposals");
-    GOVERNANCE.with_borrow(|governance| governance.list_proposals(&caller(), &req.into()))
+    with_governance(|governance| governance.list_proposals(&caller(), &req.into()))
 }
 
 #[query]
@@ -1007,166 +762,6 @@ fn http_request(request: HttpRequest) -> HttpResponse {
         "/metrics" => serve_metrics(|encoder| encode_metrics(governance(), encoder)),
         _ => HttpResponseBuilder::not_found().build(),
     }
-}
-
-// Processes the payload received and transforms it into a form the intended canister expects.
-// The arguments `proposal_id` is used by AddSnsWasm proposals.
-// `_proposal_timestamp_seconds` will be used in the future by subnet rental NNS proposals.
-fn get_effective_payload(
-    mt: gov_pb::NnsFunction,
-    payload: Vec<u8>,
-    proposal_id: u64,
-    proposal_timestamp_seconds: u64,
-) -> Result<Vec<u8>, gov_pb::GovernanceError> {
-    use gov_pb::{governance_error::ErrorType, GovernanceError, NnsFunction};
-
-    const BITCOIN_SET_CONFIG_METHOD_NAME: &str = "set_config";
-    const BITCOIN_MAINNET_CANISTER_ID: &str = "ghsi2-tqaaa-aaaan-aaaca-cai";
-    const BITCOIN_TESTNET_CANISTER_ID: &str = "g4xu7-jiaaa-aaaan-aaaaq-cai";
-
-    match mt {
-        NnsFunction::BitcoinSetConfig => {
-            // Decode the payload to get the network.
-            let payload = match Decode!([decoder_config()]; &payload, BitcoinSetConfigProposal) {
-              Ok(payload) => payload,
-              Err(_) => {
-                return Err(GovernanceError::new_with_message(ErrorType::InvalidProposal, "Payload must be a valid BitcoinSetConfigProposal."));
-              }
-            };
-
-            // Convert it to a call canister payload.
-            let canister_id = CanisterId::from_str(match payload.network {
-                BitcoinNetwork::Mainnet => BITCOIN_MAINNET_CANISTER_ID,
-                BitcoinNetwork::Testnet => BITCOIN_TESTNET_CANISTER_ID,
-            }).expect("bitcoin canister id must be valid.");
-
-            let encoded_payload = Encode!(&CallCanisterProposal {
-                canister_id,
-                method_name: BITCOIN_SET_CONFIG_METHOD_NAME.to_string(),
-                payload: payload.payload
-            })
-            .unwrap();
-
-            Ok(encoded_payload)
-        }
-        NnsFunction::SubnetRentalRequest => {
-            // Decode the payload to `SubnetRentalRequest`.
-            let payload = match Decode!([decoder_config()]; &payload, SubnetRentalRequest) {
-              Ok(payload) => payload,
-              Err(_) => {
-                return Err(GovernanceError::new_with_message(ErrorType::InvalidProposal, "Payload must be a valid SubnetRentalRequest."));
-              }
-            };
-
-            // Convert the payload to `SubnetRentalProposalPayload`.
-            let SubnetRentalRequest {
-                user,
-                rental_condition_id,
-            } = payload;
-            let proposal_creation_time_seconds = proposal_timestamp_seconds;
-            let encoded_payload = Encode!(&SubnetRentalProposalPayload {
-                user,
-                rental_condition_id,
-                proposal_id,
-                proposal_creation_time_seconds,
-            }).unwrap();
-
-            Ok(encoded_payload)
-        }
-
-        | NnsFunction::AddSnsWasm => {
-            let payload = add_proposal_id_to_add_wasm_request(&payload, proposal_id)?;
-
-            Ok(payload)
-        }
-
-        // NOTE: Methods are listed explicitly as opposed to using the `_` wildcard so
-        // that adding a new function causes a compile error here, ensuring that the developer
-        // makes an explicit decision on how the payload is handled.
-        NnsFunction::Unspecified
-        | NnsFunction::UpdateElectedHostosVersions
-        | NnsFunction::UpdateNodesHostosVersion
-        | NnsFunction::ReviseElectedHostosVersions
-        | NnsFunction::DeployHostosToSomeNodes
-        | NnsFunction::AssignNoid
-        | NnsFunction::CreateSubnet
-        | NnsFunction::AddNodeToSubnet
-        | NnsFunction::RemoveNodesFromSubnet
-        | NnsFunction::ChangeSubnetMembership
-        | NnsFunction::NnsCanisterInstall
-        | NnsFunction::NnsCanisterUpgrade
-        | NnsFunction::NnsRootUpgrade
-        | NnsFunction::HardResetNnsRootToVersion
-        | NnsFunction::RecoverSubnet
-        | NnsFunction::BlessReplicaVersion
-        | NnsFunction::RetireReplicaVersion
-        | NnsFunction::ReviseElectedGuestosVersions
-        | NnsFunction::UpdateNodeOperatorConfig
-        | NnsFunction::DeployGuestosToAllSubnetNodes
-        | NnsFunction::UpdateConfigOfSubnet
-        | NnsFunction::IcpXdrConversionRate
-        | NnsFunction::ClearProvisionalWhitelist
-        | NnsFunction::SetAuthorizedSubnetworks
-        | NnsFunction::SetFirewallConfig
-        | NnsFunction::AddFirewallRules
-        | NnsFunction::RemoveFirewallRules
-        | NnsFunction::UpdateFirewallRules
-        | NnsFunction::StopOrStartNnsCanister
-        | NnsFunction::RemoveNodes
-        | NnsFunction::UninstallCode
-        | NnsFunction::UpdateNodeRewardsTable
-        | NnsFunction::AddOrRemoveDataCenters
-        | NnsFunction::UpdateUnassignedNodesConfig // obsolete
-        | NnsFunction::RemoveNodeOperators
-        | NnsFunction::RerouteCanisterRanges
-        | NnsFunction::PrepareCanisterMigration
-        | NnsFunction::CompleteCanisterMigration
-        | NnsFunction::UpdateSubnetType
-        | NnsFunction::ChangeSubnetTypeAssignment
-        | NnsFunction::UpdateAllowedPrincipals
-        | NnsFunction::UpdateSnsWasmSnsSubnetIds
-        | NnsFunction::InsertSnsWasmUpgradePathEntries
-        | NnsFunction::AddApiBoundaryNodes
-        | NnsFunction::RemoveApiBoundaryNodes
-        | NnsFunction::UpdateApiBoundaryNodesVersion // obsolete
-        | NnsFunction::DeployGuestosToAllUnassignedNodes
-        | NnsFunction::UpdateSshReadonlyAccessForAllUnassignedNodes
-        | NnsFunction::DeployGuestosToSomeApiBoundaryNodes => Ok(payload),
-    }
-}
-
-fn add_proposal_id_to_add_wasm_request(
-    payload: &[u8],
-    proposal_id: u64,
-) -> Result<Vec<u8>, GovernanceError> {
-    let add_wasm_request = match Decode!([decoder_config()]; payload, AddWasmRequest) {
-        Ok(add_wasm_request) => add_wasm_request,
-        Err(e) => {
-            return Err(GovernanceError::new_with_message(
-                ErrorType::InvalidProposal,
-                format!("Payload must be a valid AddWasmRequest. Error: {e}"),
-            ));
-        }
-    };
-
-    let wasm = add_wasm_request
-        .wasm
-        .ok_or(GovernanceError::new_with_message(
-            ErrorType::InvalidProposal,
-            "Payload must contain a wasm.",
-        ))?;
-
-    let add_wasm_request = AddWasmRequest {
-        wasm: Some(SnsWasm {
-            proposal_id: Some(proposal_id),
-            ..wasm
-        }),
-        ..add_wasm_request
-    };
-
-    let payload = Encode!(&add_wasm_request).unwrap();
-
-    Ok(payload)
 }
 
 fn main() {
