+++ rs/nns/governance/canister/tests.rs
@@ -38,67 +38,3 @@ fn test_implemented_interface_matches_declared_interface_exactly() {
     let result = service_equal(declared_interface, implemented_interface);
     assert!(result.is_ok(), "{:?}\n\n", result.unwrap_err());
 }
-
-#[test]
-fn test_set_time_warp() {
-    let mut environment = CanisterEnv::new();
-
-    let start = environment.now();
-    environment.set_time_warp(GovTimeWarp { delta_s: 1_000 });
-    let delta_s = environment.now() - start;
-
-    assert!(delta_s >= 1000, "delta_s = {}", delta_s);
-    assert!(delta_s < 1005, "delta_s = {}", delta_s);
-}
-
-#[test]
-fn test_get_effective_payload_sets_proposal_id_for_add_wasm() {
-    let mt = gov_pb::NnsFunction::AddSnsWasm;
-    let proposal_id = 42;
-    let wasm = vec![1, 2, 3];
-    let canister_type = 3;
-    let hash = vec![1, 2, 3, 4];
-    let payload = Encode!(&AddWasmRequest {
-        wasm: Some(SnsWasm {
-            proposal_id: None,
-            wasm: wasm.clone(),
-            canister_type,
-        }),
-        hash: hash.clone(),
-    })
-    .unwrap();
-
-    let effective_payload = get_effective_payload(mt, payload, proposal_id, 0).unwrap();
-
-    let decoded = Decode!(&effective_payload, AddWasmRequest).unwrap();
-    assert_eq!(
-        decoded,
-        AddWasmRequest {
-            wasm: Some(SnsWasm {
-                proposal_id: Some(proposal_id), // The proposal_id should be set
-                wasm,
-                canister_type
-            }),
-            hash
-        }
-    );
-}
-
-#[test]
-fn test_get_effective_payload_overrides_proposal_id_for_add_wasm() {
-    let mt = gov_pb::NnsFunction::AddSnsWasm;
-    let proposal_id = 42;
-    let payload = Encode!(&AddWasmRequest {
-        wasm: Some(SnsWasm {
-            proposal_id: Some(proposal_id - 1),
-            ..SnsWasm::default()
-        }),
-        ..AddWasmRequest::default()
-    })
-    .unwrap();
-
-    let effective_payload = get_effective_payload(mt, payload, proposal_id, 0).unwrap();
-
-    let decoded = Decode!(&effective_payload, AddWasmRequest).unwrap();
-    assert_eq!(decoded.wasm.unwrap().proposal_id.unwrap(), proposal_id);
-}
